# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Adhams ERP & Logistics Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /orders:
    get:
      operationId: listOrders
      tags: [orders]
      summary: List orders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, confirmed, reserved, dispatched, delivered, cancelled]
        - name: dealer_id
          in: query
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Order list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OrderListResponse"
    post:
      operationId: createOrder
      tags: [orders]
      summary: Create order
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrder"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Order"
...
  /dealers:
    get:
      operationId: listDealers
      tags: [dealers]
      summary: List dealers
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: Dealer list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DealerListResponse"
    post:
      operationId: createDealer
      tags: [dealers]
      summary: Create dealer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateDealer"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Dealer"
...
components:
  schemas:
...
    Order:
      type: object
      properties:
        id:
          type: integer
        orderNumber:
          type: string
        dealerId:
          type: integer
        dealerName:
          type: string
        status:
          type: string
          enum: [pending, confirmed, reserved, dispatched, delivered, cancelled]
        totalAmount:
          type: number
        advancePaid:
          type: number
        balanceAmount:
          type: number
        items:
          type: array
          items:
            $ref: "#/components/schemas/OrderItem"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - orderNumber
        - dealerId
        - dealerName
        - status
        - totalAmount
        - advancePaid
        - balanceAmount
        - items
        - createdAt
        - updatedAt
    OrderItem:
      type: object
      properties:
        productId:
          type: integer
        productName:
          type: string
        quantity:
          type: integer
        unitPrice:
          type: number
        totalPrice:
          type: number
      required:
        - productId
        - productName
        - quantity
        - unitPrice
        - totalPrice
    OrderListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Order"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    CreateOrder:
      type: object
      properties:
        dealerId:
          type: integer
        items:
          type: array
          items:
            type: object
            properties:
              productId:
                type: integer
              quantity:
                type: integer
              unitPrice:
                type: number
            required:
              - productId
              - quantity
              - unitPrice
        advancePaid:
          type: number
      required:
        - dealerId
        - items
...
    Dealer:
      type: object
      properties:
        id:
          type: integer
        dealerCode:
          type: string
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        city:
          type: string
        state:
          type: string
        gstNumber:
          type: string
        creditLimit:
          type: number
        outstandingBalance:
          type: number
        isActive:
          type: boolean
        commissionSlab:
          type: string
        createdAt:
          type: string
      required:
        - id
        - dealerCode
        - name
        - contactPerson
        - phone
        - city
        - state
        - isActive
        - createdAt
    DealerListResponse:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: "#/components/schemas/Dealer"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - items
        - total
        - page
        - limit
    CreateDealer:
      type: object
      properties:
        name:
          type: string
        contactPerson:
          type: string
        phone:
          type: string
        email:
          type: string
        city:
          type: string
        state:
          type: string
        gstNumber:
          type: string
        creditLimit:
          type: number
        commissionSlab:
          type: string
      required:
        - name
        - contactPerson
        - phone
        - city
        - state
...